requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-camel-case==0.2.0
get==1.0.3
gunicorn==19.9.0
idna==2.7
Pillow==5.3.0
pipenv==2018.10.13
post==1.0.2
psycopg2-binary==2.7.5
public==2.0.1
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
query-string==1.0.2
request==1.0.2
six==1.11.0
virtualenv==16.0.0
virtualenv-clone==0.4.0
whitenoise==4.0
